requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core
langchain-text-splitters
langchain-ollama>=0.1.0
langchain-chroma
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langgraph>=0.0.20
langsmith>=0.0.70
pypdf2>=3.0.0
python-docx>=1.0.0
unstructured>=0.10.0
sentence-transformers>=2.0.0
openai>=1.0.0
groq>=0.4.0
google-generativeai>=0.7.0
requests>=2.30.0
aiofiles>=23.0.0
redis>=5.0.0
celery>=5.3.0
tiktoken>=0.5.0